+++ include/modules/network.hpp
@@ -55,6 +55,7 @@ namespace modules {
     int m_udspeed_minwidth{0};
     bool m_accumulate{false};
     bool m_unknown_up{false};
+    bool m_metric_units{true};
     string m_udspeed_unit{"B/s"};
   };
 }  // namespace modules